Implement initial battery consuming logic for drone

+++ EscapeRoom/Assets/BatterySystem/BatteryClass.cs
@@ -31,6 +31,6 @@ public class BatteryClass : MonoBehaviour
 
     public float GetBatteryPercent()
     {
-        return (Charge / MaxCharge) * 100;
+        return (Charge / (float)MaxCharge) * 100;
     }
 }
